pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ibkr-mcp"
version = "0.1.0"
description = "MCP server for Interactive Brokers Gateway — 6 tools for market data and account access via Claude"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "ib_async",
    "nest_asyncio",
    "pandas",
    "pyarrow",
    "pyyaml",
    "python-dotenv",
    "fastmcp",
    "certifi",
    "zstandard",
]

[project.scripts]
ibkr-mcp = "ibkr.server:main"

[tool.hatch.build.targets.wheel]
packages = ["ibkr"]

[tool.hatch.build.targets.sdist]
include = ["ibkr/", "pyproject.toml", "README.md", "LICENSE"]
